+++ psychopy/demos/coder/iohub/serial/customparser.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
 """
 Demo using a custom serial rx parser to generate ioHub Serial Device events.
 
@@ -8,7 +11,8 @@ custom parser function that the ioHub Serial device uses during runtime.
 serial device receives back as rx data. **
 
 """
-from __future__ import print_function
+from __future__ import absolute_import, division, print_function
+
 import time
 from psychopy import core, visual
 from psychopy.iohub import launchHubServer
@@ -36,7 +40,7 @@ event_parser_info = dict(parser_function="parseserial.checkForSerialEvents",
                          parser_kwargs=dict(var1='not used', var2=1234))
 # configure iohub
 exp_code = 'serial_demo'
-sess_code = 'S_{0}'.format(long(time.mktime(time.localtime())))
+sess_code = 'S_{0}'.format(int(time.mktime(time.localtime())))
 iohubkwargs = {'experiment_code': exp_code,
                'session_code': sess_code,
                'serial.Serial': dict(name='serial',
